Replace the run carousel with a stacked route list
Two path cards do not need snap scrolling or prev/next. Each route
now sits in the page column with its count, day span, and last run.
